+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Exclusive support fo `laravel-datatables:8.*`.
 - Change namespace from `Yajra\Datatables` to `Yajra\DataTables`.
 - Force implementation/upgrade of `dataTable()` method.
+- Use `snappy` as default pdf generator.
 
 ### v2.0.2 - 06-30-2017
 - Fix min php requirements.